Implement a builder for the Radiotap struct
All the fields of the Radiotap struct are public, and the struct
already implements Default, so while it *is* possible to build values
in code, doing so is inconvenient: the Header value has to be populated
correctly based on which fields are set and in which order they appear
in the serialized format. A dedicated builder struct allows to do it
fluently and prevent potential mistakes.

This requires minimal changes to the existing API, most of which are
optional:

* Kind::bit() offers a conversion from an enum variant back to a bit
number value; that value defines the order in which fields appear
in the serialized format

* Radiotap::build() is a helper function that starts the build process

* Implementing Default makes passing flag values slightly less verbose

The primary motivation for this change is to allow building Radiotap
values programmatically and later serialize them to the wire format.