Implement unparsing of Radiotap values
Add a new method called `unparse()` that serializes a Radiotap value
into a stream of bytes according to the rules of the Radiotap wire
format. In conjunction with the `build()` method this allows to build
Radiotap values programmatically, e.g. for packet injection.

Adding a new test uncovered a few issues in the existing code related
to the precedence of bitwise operations that are now fixed.